Refactor to ES modules with Vite bundling
