+++ index.html
@@ -58,6 +58,6 @@
     </div>
   </div>
 
-  <script src="script.js" defer></script>
-</body>
+    <script type="module" src="dist/main.js"></script>
+  </body>
 </html>
