+++ src/ui/error.js
@@ -1,0 +1,33 @@
+import { escapeHTML, nowStamp, formatDuration } from "../utils.js";
+import { API_BASE } from "../api.js";
+
+export function renderLoading(address) {
+  document.getElementById("result").innerHTML = `
+    <div class="card">
+      <div class="card__header">
+        <h2 class="card__title">Looking up demographics…</h2>
+        <span class="updated">Started ${nowStamp()}</span>
+      </div>
+      ${address ? `<p class="note">Address: <strong>${escapeHTML(address)}</strong></p>` : ""}
+      <div class="callout">Fetching county, languages, English proficiency, population, income, DAC, and alerts…</div>
+      <p class="note">Elapsed: <span id="searchTimer">0m 00s</span></p>
+    </div>
+  `;
+}
+
+export function renderError(message, address, elapsedMs) {
+  document.getElementById("result").innerHTML = `
+    <div class="card" role="alert">
+      <div class="card__header">
+        <h2 class="card__title">Unable to retrieve data</h2>
+        <span class="updated">${nowStamp()}</span>
+      </div>
+      ${address ? `<p class="note">Address: <strong>${escapeHTML(address)}</strong></p>` : ""}
+      <div class="callout" style="border-left-color:#b45309;">
+        ${escapeHTML(message || "Please try again with a different address.")}
+      </div>
+      <p class="note">Search took ${formatDuration(elapsedMs)}.</p>
+      <p class="note">API base: <code>${escapeHTML(API_BASE)}</code>. If your API has a prefix, adjust <code>API_PATH</code>.</p>
+    </div>
+  `;
+}
