+++ vite.config.js
@@ -1,0 +1,17 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  build: {
+    rollupOptions: {
+      input: 'src/main.js',
+      output: {
+        entryFileNames: 'main.js',
+        chunkFileNames: '[name].js',
+        manualChunks: {
+          maps: ['src/maps.js'],
+          pdf: ['src/pdf.js'],
+        },
+      },
+    },
+  },
+});
